requirements.txt
huggingface-hub>=0.23.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
fastapi>=0.110.0
uvicorn>=0.29.0
